Move TestAddDocument and related helpers
down from AbstractUserDiagnosticTest to
AbstractCodeActionOrUserDiagnosticTest.

CodeAction based refactorings that add new documents,
that are not diagnostic based, such as MoveTypeToFile, need these too.
